Use `UInt32` for `gitCommitParentID()` index parameter
Changed the `n` parameter's type from `UInt` to `UInt32` to remove the
type cast that could potentially overflow on some platforms. Although an
overflow is unlikely in practice, the caller should handle conversions.

+++ Sources/SwiftLibgit2/Commit/Commit-Functions.swift
@@ -510,12 +510,12 @@ public func gitCommitParent(
 /// [`git_commit_parent_id()`](https://libgit2.org/docs/reference/main/commit/git_commit_parent_id.html)
 public func gitCommitParentID(
     commit  : OpaquePointer,
-    n       : UInt
+    n       : UInt32
 ) -> GitOID
 {
     let parentCommitOID: UnsafePointer<git_oid> = git_commit_parent_id(
         commit,
-        UInt32(n)
+        n
     )
     
     return GitOID(cValue: parentCommitOID.pointee)
